Minor bugfixes
git-svn-id: http://nexuse2e-server.svn.sourceforge.net/svnroot/nexuse2e-server/trunk@331 0f8715a3-b92c-0410-9248-d6441137b268

+++ src/main/java/org/nexuse2e/ui/action/partners/PartnerDeleteAction.java
@@ -62,7 +62,9 @@ public class PartnerDeleteAction extends NexusE2EAction {
         try {
             PartnerPojo partner = Engine.getInstance().getActiveConfigurationAccessService().getPartnerByNxPartnerId(
                     form.getNxPartnerId() );
-            Engine.getInstance().getActiveConfigurationAccessService().deletePartner( partner );
+            if (partner != null) {
+                Engine.getInstance().getActiveConfigurationAccessService().deletePartner( partner );
+            }
         } catch ( ReferencedPartnerException e ) {
             for (ParticipantPojo participant : e.getReferringObjects()) {
                 ActionMessage errorMessage = new ActionMessage(
